add animation to hero Fixes #69
