+++ src/components/sticky-title/sticky-title.tsx
@@ -2,14 +2,13 @@ import { Blocks } from "@/components/blocks";
 import type { StickyTitleType } from "@/payload-types";
 import { cva } from "class-variance-authority";
 import styles from "./sticky-title.module.css";
-import { AnimatedTitle } from "./elements/animated-title";
 
 const stickyTitleStyles = cva(styles.container);
 
 function StickyTitle(props: StickyTitleType) {
 	return (
 		<section className={stickyTitleStyles()}>
-			<AnimatedTitle>{props.title}</AnimatedTitle>
+			<h2 className={styles.title}>{props.title}</h2>
 			<div>
 				<Blocks blocks={props.blocks} />
 			</div>
